split dsl from AT module
