+++ acceptance-tests/build.gradle
@@ -1,44 +1,3 @@
 
-apply plugin: 'com.adarshr.test-logger'
 
 jar { enabled = false }
-
-dependencies {
-  testImplementation project(":octopus-sdk")
-  testImplementation project(":octopus-openapi-wrapper")
-  testImplementation 'org.testcontainers:testcontainers'
-  testImplementation 'com.squareup.okhttp3:okhttp'
-  testImplementation 'com.google.code.gson:gson'
-  testImplementation 'com.google.guava:guava'
-  testImplementation 'org.apache.logging.log4j:log4j-api'
-  testImplementation 'org.apache.logging.log4j:log4j-core'
-  testImplementation 'org.apache.logging.log4j:log4j-slf4j-impl'
-
-  testImplementation 'org.junit.jupiter:junit-jupiter-api'
-  testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-  testImplementation 'org.junit.jupiter:junit-jupiter-params'
-  testImplementation 'org.assertj:assertj-core'
-
-  testImplementation 'org.mock-server:mockserver-netty'
-  testImplementation 'org.mock-server:mockserver-client-java'
-  testImplementation 'org.assertj:assertj-core'
-  testImplementation 'org.testcontainers:testcontainers'
-  testImplementation 'org.testcontainers:junit-jupiter'
-}
-
-test.enabled = false
-
-task acceptanceTest(type: Test) {
-  java {
-    toolchain {
-      languageVersion = JavaLanguageVersion.of(11)
-    }
-  }
-  mustRunAfter rootProject.subprojects*.test
-  description = 'Runs the Octopus Java SDK acceptance tests.'
-  group = 'verification'
-  useJUnitPlatform()
-  outputs.upToDateWhen { false }
-
-  testLogging.showStandardStreams = false;
-}
