--- acceptance-tests/src/test/resources/log4j2.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="DEBUG" monitorInterval="30">
-  <Properties>
-    <Property name="root.log.level">INFO</Property>
-  </Properties>
-  <Appenders>
-    <Console name="Console" target="SYSTEM_OUT">
-      <PatternLayout pattern="%d{HH:mm:ss.SSS} | %t | %-5level | %c{1} | %msg%n" />
-    </Console>
-  </Appenders>
-  <Loggers>
-    <Root level="${sys:root.log.level}">
-      <AppenderRef ref="Console" />
-    </Root>
-  </Loggers>
-</Configuration>
