+++ acceptance-tests/tests/src/test/java/com/octopus/sdk/test/BuildInformationAcceptanceTest.java
