+++ build.gradle
@@ -48,7 +48,6 @@ def buildAliases = ['dev': [
 
 subprojects {
     apply plugin: 'java'
-    apply plugin: 'maven-publish'
     apply plugin: 'io.spring.dependency-management'
     apply plugin: 'net.ltgt.errorprone'
     apply from: "${rootDir}/gradle/versions.gradle"
@@ -153,7 +152,8 @@ subprojects {
         testLogging.showStandardStreams = false
     }
 
-    if (project.name != "acceptance-tests") {
+    // offer to publish only projects which have a main sourceset containing files
+    if (project.sourceSets.getByName("main").allJava.size() != 0) {
         apply plugin: "com.jfrog.artifactory"
         apply plugin: 'maven-publish'
