+++ settings.gradle
@@ -2,4 +2,6 @@ rootProject.name='octopus-java-sdk'
 include 'octopus-sdk'
 include 'octopus-openapi-wrapper'
 include 'acceptance-tests'
+include 'acceptance-tests:dsl'
+include 'acceptance-tests:tests'
 include 'octopus-openapi-wrapper'
